+++ frontend/src/components/browser/toolbars/top/filter-by-select.vue
@@ -121,6 +121,15 @@ export default {
             names.push(key);
           }
         }
+        // Add selected filters to choices if they aren't there already.
+        if (this.filters?.length) {
+          for (const [filterKey, filterValue] of Object.entries(this.filters)) {
+            if (filterValue?.length && !names.includes(filterKey)) {
+              names.push(filterKey);
+            }
+          }
+          names.sort();
+        }
         return names;
       },
     }),